requirement.txt
selenium
pytest
pytest-rerunfailures
pytest-xdist
allure-pytest
PyYAML
pytest-ordering
request